Prepare model subclassing: move methods into model.py
